refactor: 네이밍 개선 및 불필요한 코드 삭제
